skapat en ny klass customerhandler med metod CreateNewCustomer(), den gör allt "SaveCustomerToFile()" gjorde.. lite clean code!
